Background: simplify interface and implementation of the module
Instead of having 'struct Background' describe a collection of pointers
related to each of the four GBA backgrounds, the structure now describes
a logical background characterized by its attributes. This removes the
need for 'struct background_Config' to exist.

The list of registers has now been distributed among the inline
functions ('background_config' and 'background_set_offset'), so that
they do not need to be stored in a structure.