Timer: move implementation to the header file
The reason for this change is the same as the one for the background
driver: C is much more readable than assembly. While it is true that
making these functions inline clutters the header file, they are small
enough.

Instead of relying on header files being easy to read, documentation
(i.e. man pages) should be used instead.