Random: rename, improve and move to header the two functions
The 'random' function now accepts a 'bound' parameter. When calculating
the output, the 'modulo' operator is replaced by one multiplication and
one division (by a power of 2), making the operation much faster for
bounds that are not powers of 2.

The 'random_seed' function no longer accepts a 'reset' parameter.
Instead, it now returns the previously stored seed.