Display: use (x, y, width, height) to set a window's viewport
Since the hardware registers of windows 0 and 1 store the coordinates
(x0, y0, x0 + 1, y0 + 1) it makes more sense to pass (x, y, width,
height) and calculate those coordinates inside the function.

Also, the name of the function was changed to 'display_window_viewport',
since its role is to set not just its size, but the entire viewport.