Add 'backup' module
This new module offers functions for using the cartridge's backup chip
(SRAM, FLASH ROM or EEPROM).

Data reads from these chips are not guaranteed to be made from WRAM (as
recommended by documentation), but this should not be a problem for new
cartridges and emulators. If necessary, the implementation could be
changed to follow that recommendation without changing the interface.

EEPROM support has not yet been added.