mGBA: add 'mgba_printf' function
This function is a much simplified version of 'printf'. It makes it
possible to make formatted output conversions, which can be very useful
for printing certain values while debugging.

Since this is a variadic function, the standard header 'stdarg.h' is
required, therefore the '-nostdinc' GCC option has been removed.