Display: make 'display_config' only take Video Mode as input
The rest of the configuration (enable backgrounds, sprites etc...) is
either already determined or relegated to other modules.

Almost all display-related settings are reset: backgrounds and windows
are disabled, and raster page 0 is selected. Only the 'force blank' bit
remains unmodified.

Some settings have become inaccessible to the user: access to OAM during
HBlank is disabled, sprite mapping is set to 'linear' and sprites are
always enabled.

If configuring a bitmap Video Mode, background 2 is enabled, meaning
that it's not necessary to call 'background_toggle'.