Audio: move handling of special channel IDs to header file
The 'stop', 'loop', 'volume' and 'panning' functions apply their
operation to all channels if a negative channel ID is given, while also
ignoring other invalid channel IDs. This behavior should be the same,
regardless of the chosen driver.

Bringing that code to the header file prevents code duplication, and
allows the compiler to apply loop unrolling and constant propagation.