Math: add LUT-based implementation of Tangent function (math_tan)
Just like math_sin and math_cos, this function returns a fixed-point
result with 1=0x4000. The results are fairly accurate, with the lowest
accuracy near points in which the tangent is not defined.

WIP: the LUT is missing the last entry, which theoretically should be
infinity. A good value should be found so that the results are as
accurate as possible.