Audio (mixer): optimize sample retrieval in '_mixer_update'
Instead of keeping the temporary value of position a 12-bit value, keep
it as a 32-bit value that can go above 0x1000. By doing this, the 'bic'
instruction in the 'samples loop' can be removed, improving performance.

But also, now (position / 0x1000) cannot be added to the data pointer
directly. A temporary value 'data_plus_pos' must be used, calculated
using 'data' as a base pointer and (position / 0x1000) as offset.