Math: try to shrink bounds of binary search in 'math_sqrt'
Previously, the argument was used as higher bound of the binary search,
but it's possible to set a more restrictive bound by looking at powers
of two. Also, the new code does not need to handle any special cases.

For arguments higher than U16_MAX, it would be faster to skip the bounds
shrinking phase, but for the moment I prefer to avoid adding more
complexity to the function.