Move to a single-header include system
Instead of including a header for each library module required, it is
easier to just include one single header (libsimplegba.h). This approach
is also used by established libraries like SDL.

The main downside of this is the increased compilation time, caused by
all headers being included even if they are not required.

The 'libsimplegba.h' header should NOT be used within the library to
avoid issues with recursive dependencies.