Doc (libsimplegba): shorten module descriptions + add 'Usage' subsection
Since modules should not be included individually anymore, their header
files should not be listed in manuals. In the list, header names are
replaced by a very short description of the module.

The longer descriptions are removed, suggesting to the user to instead
read the module manuals.

The 'Usage' subsection briefly explains how to setup a Makefile to
include the library and that the 'AgbMain' function needs to be defined.